requirements.txt
# ============================================================================
# PHARMA SEARCH - CONSOLIDATED REQUIREMENTS
# Python package dependencies with latest versions as of 2025
# ============================================================================

# Core Web Framework
fastapi==0.115.6
uvicorn[standard]==0.32.1
pydantic==2.11.7
pydantic-settings==2.8.0

# Database
asyncpg==0.30.0
python-dotenv==1.0.1

# Data Processing & Text Processing
rapidfuzz==3.10.1
transliterate==1.10.2
unidecode==1.3.8
tqdm==4.67.1

# Machine Learning & AI
torch>=2.5.0
torchvision>=0.20.0
torchaudio>=2.5.0
transformers==4.53.2
sentence-transformers==3.3.1
scikit-learn==1.7.1
numpy==2.2.0

# Web Scraping & Browser Automation (for scrapers - optional)
requests==2.32.3
beautifulsoup4==4.12.3

# Email & Communications
aiosmtplib==3.0.2

# Development & Testing
pytest==8.3.4
pytest-asyncio==0.25.0

# Data Formats
openpyxl==3.1.5

# Logging
structlog==24.4.0

# ============================================================================
# OPTIONAL CUDA DEPENDENCIES
# Install these separately based on your hardware:
# 
# For CUDA 11.8+:
#   pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu118
#
# For CPU only:
#   pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cpu
#
# For Apple Silicon (Metal):
#   pip install torch torchvision torchaudio
# ============================================================================

# ============================================================================
# INSTALLATION NOTES
# ============================================================================
#
# 1. Install base dependencies:
#    pip install -r requirements.txt
#
# 2. For ML features, run the setup script:
#    python backend/scripts/setup_ml.py
#
# 3. For scrapers (Node.js dependencies):
#    cd frontend && npm install
#
# 4. Database setup:
#    - Set DATABASE_URL environment variable
#    - Run: psql -d your_database < schema.sql
#    - Run: psql -d your_database < seed.sql
#
# ============================================================================